move getSubmission
