+++ src/app/submissions/[sid]/RunInfo.tsx
@@ -1,8 +1,8 @@
-import type { SubmissionInfo } from "@utils/api/getSubmissionInfo";
 import { getVerdictTextClass } from "@utils/shared";
 import { Suspense } from "react";
 
 import { Heading } from "@/components";
+import type { SubmissionInfo } from "@/features/submissions";
 import type { Results } from "@/types";
 import { RESULTS_DICT } from "@/types";
