+++ src/features/problems/queries/getProblem.ts
@@ -3,7 +3,7 @@ import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 import { parseJudge } from "@utils/shared";
 import { unstable_cache } from "next/cache";
 
-import type { DetailedProblemInfo } from "../types";
+import type { DetailedProblemInfo } from "@/features/problems";
 
 export const getProblem = async (code: string) => {
   try {
